+++ bzt/__init__.py
@@ -17,7 +17,7 @@ import os
 import sys
 import signal
 
-VERSION = "0.2.17"
+VERSION = "0.2.18"
 
 
 def signal_handler(sig, frame):